+++ matchms/similarity/__init__.py
@@ -19,6 +19,7 @@ It is also easily possible to add own custom similarity measures or import exter
 (such as `Spec2Vec <https://github.com/iomega/spec2vec>`_).
 """
 from .BinnedEmbeddingSimilarity import BinnedEmbeddingSimilarity
+from .BlinkCosine import BlinkCosine
 from .CosineGreedy import CosineGreedy
 from .CosineHungarian import CosineHungarian
 from .FingerprintSimilarity import FingerprintSimilarity
@@ -33,6 +34,7 @@ from .PrecursorMzMatch import PrecursorMzMatch
 
 __all__ = [
     "BinnedEmbeddingSimilarity",
+    "BlinkCosine",
     "CosineGreedy",
     "CosineHungarian",
     "FingerprintSimilarity",
@@ -56,9 +58,11 @@ def get_similarity_function_by_name(similarity_function_name: str):
         Name of the similarity function.
     """
     names = __all__
-    functions = [BinnedEmbeddingSimilarity, CosineGreedy, CosineHungarian, FingerprintSimilarity,
-                 FlashSimilarity, IntersectMz, MetadataMatch, ModifiedCosine, NeutralLossesCosine,
-                 ParentMassMatch, PrecursorMzMatch]
+    functions = [BinnedEmbeddingSimilarity, BlinkCosine,
+                 CosineGreedy, CosineHungarian,
+                 FingerprintSimilarity, FlashSimilarity,
+                 IntersectMz, MetadataMatch, ModifiedCosine,
+                 NeutralLossesCosine, ParentMassMatch, PrecursorMzMatch]
 
     assert similarity_function_name in names, f"Unknown similarity function: {similarity_function_name}"
     assert len(names) == len(functions), "Number of similarity functions and names do not match"